+++ test/panda/core/RecvTest2.java
@@ -13,7 +13,7 @@ public class RecvTest2
 	public static void main(String[] args) throws Exception
 	{
 		short topicID = Short.valueOf(args[0]).shortValue();
-		TopicInfo topicInfo = new TopicInfo("239.9.9.9", Integer.valueOf(9001), Short.valueOf(topicID), "TEST_TOPIC");
+		TopicInfo topicInfo = new TopicInfo("239.9.9.9", Integer.valueOf(9001), Integer.valueOf(topicID), "TEST_TOPIC");
 
 		ReceiverTester rt = new ReceiverTester();
 		rt.SubscribeToTest(0, topicInfo);
